Separate cli from core, fix end-to-end test, and fix scripts
