+++ package.json
@@ -19,13 +19,13 @@
   },
   "main": "lib/index.js",
   "bin": {
-    "cpf": "./lib/index.js"
+    "cpf": "./lib/cli/index.js"
   },
   "scripts": {
     "test": "jest",
-    "start": "node lib/index.js",
+    "start": "node lib/cli/index.js",
     "build": "yarn babel src --extensions \".js,.ts\" --out-dir lib --copy-files --no-copy-ignored",
-    "dev": "ts-node -r tsconfig-paths/register src/index.ts --transpile-only --ignore-watch node_modules --no-notify",
+    "dev": "ts-node -r tsconfig-paths/register src/cli/index.ts",
     "prepare": "yarn clean:lib && yarn build",
     "install:windows": "yarn build && yarn global add file:%INIT_CWD%",
     "install:unix": "yarn build && yarn global add file:$INIT_CWD",
