+++ src/cli/commands/index.ts
@@ -1,0 +1,1 @@
+export { default as CreateCommand } from './CreateCommand'
