+++ src/core/commands/CreateCommand/DefaultOptions.ts
@@ -1,0 +1,8 @@
+export const DefaultOptions = {
+  name: '',
+  templatesFolder: '__file-templates__',
+  encoding: 'utf-8',
+  replaceNames: undefined,
+  replaceContent: undefined,
+  brackets: true,
+}
