+++ src/core/commands/CreateCommand/index.ts
@@ -1,0 +1,2 @@
+export * from './CreateCommand'
+export * from './DefaultOptions'
