+++ src/core/errors/NotFoundError.ts
