+++ src/core/errors/SyntaxError.ts
