+++ src/core/errors/index.ts
