+++ src/core/locales/EN_US.json
