+++ src/core/locales/index.ts
