+++ src/core/utils/TextUtils.ts
